requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0
requests==2.31.0
paramiko==3.4.0
docker==7.0.0
psutil==5.9.6
gunicorn==21.2.0
python-dotenv==1.0.0
schedule==1.2.1
click==8.1.7
Werkzeug==3.0.1
eventlet==0.33.3
